requirements.txt
pillow
numpy
matplotlib
scipy
scikit-learn
imagehash
urllib3
requests
pandas
tflite-runtime; sys_platform == "linux"
tensorflow-macos
